# bpo-37295: Speed up math.comb(n, k) for 0 <= k <= n <= 67