--- a/Misc/NEWS.d/next/Library/2021-12-27-15-52-28.bpo-37295.s3LPo0.rst
+++ b/Misc/NEWS.d/next/Library/2021-12-27-15-52-28.bpo-37295.s3LPo0.rst
@@ -1,0 +1,1 @@
+Add fast path for `0 <= k <= n <= 67` for :func:`math.comb`.