This PR speeds up `math.comb(n, k)` for very small `k` and `n`.

The existing tests already include exhaustive testing of correctness of `math.comb(n, k)` for `0 <= k <= n <= 100`, so I didn't add any more unit tests.

Some sample timings on my machine (comparing commit fd3a4bba87415c96f3979c8e09932aa948c8bc6c on this branch with commit 2e3e0d23adca8d83722d939d6abd1e467d7578f7 on the main branch):

`./python.exe -m timeit -s "from math import comb;n, k = 1, 1" "comb(n, k)"`: difference not significant
- main: 5000000 loops, best of 5: 35.5 nsec per loop
- here: 10000000 loops, best of 5: 34.2 nsec per loop

`./python.exe -m timeit -s "from math import comb;n, k = 5, 3" "comb(n, k)"`: ~23% faster (but may not be significant)
- main: 5000000 loops, best of 5: 38.6 nsec per loop
- here: 10000000 loops, best of 5: 31.5 nsec per loop

`./python.exe -m timeit -s "from math import comb;n, k = 67, 25" "comb(n, k)"`: ~9.7x faster
- main: 500000 loops, best of 5: 447 nsec per loop
- here: 5000000 loops, best of 5: 46.1 nsec per loop






<!-- issue-number: [bpo-37295](https://bugs.python.org/issue37295) -->
https://bugs.python.org/issue37295
<!-- /issue-number -->